Creating an ambiorix + htmx app

Creating an ambiorix + htmx app

Usage

ambhtmx_app(
  dbname = NULL,
  value = tibble::tibble(),
  protocol = NULL,
  host = NULL,
  port = NULL,
  live = "",
  favicon = NULL,
  render_index = NULL,
  render_row = NULL
)

Arguments

dbname

file path to store a SQLite database (optional).

value

a 1 row tibble with the names and types of the columns (optional)

protocol

(default AMBHTMX_PROTOCOL or http)

host

(default AMBHTMX_HOST or 127.0.0.1)

port

(default AMBHTMX_PORT or 3000)

live

script with the file path (optional)

favicon

(optional)

render_index

function to be stored as a model method (optional)

render_row

function to be stored as a model method (optional)

Value

A list with the ambiorix app, the running context and the model methods.