Render and send login page response

Render and send login page response

Usage

process_login_get(
  req,
  res,
  page_title = "Login",
  main = NULL,
  id = "login_form",
  login_url = "/login",
  style = "margin: 20px",
  cookie_errors = "errors"
)

Arguments

req

request object

res

response object

page_title

if you need to customize the title of the page

main

if you need to customize the body of the login page

id

if you need to customize the id of the login form

login_url

if you need to customize the url of the login form

style

if you need to customize the styles of the login form

cookie_errors

if you need to customize the name of the errors cookie

Value

the login page response